refactor: validations
