+++ src/lib/validations/README.md
@@ -1,0 +1,11 @@
+# Validations
+
+This folder contains the Yup validation schemas.
+
+## Files
+
+The following files are included in this folder:
+
+- `auth.validations.js`: File containing all user and auth related validations
+- `board.validations.js`: File containing all validations related to the board
+- `task.validations.js`: File containing all validations related to the a task like for label forms, comments and create and update task
