+++ src/lib/validations/board.validations.ts
@@ -1,6 +1,6 @@
 import * as Yup from 'yup';
 
-const newBoardFormSchema = Yup.object().shape({
+export const newBoardFormSchema = Yup.object().shape({
 	file: Yup.mixed(),
 	name: Yup.string()
 		.required('Please enter a board name')
@@ -8,4 +8,10 @@ const newBoardFormSchema = Yup.object().shape({
 	isPrivate: Yup.boolean().required('Please specify privacy preference'),
 	owner: Yup.string().required('Please enter the owner name'),
 });
-export default newBoardFormSchema;
+
+export const boardDescriptionFormSchema = Yup.object().shape({
+	id: Yup.string().required(),
+	description: Yup.string()
+		.required('Please enter a description')
+		.max(5000, 'Description  cannot be longer than 5000 characters'),
+});
