--- src/validations/authForm.validation.ts
@@ -1,12 +1,0 @@
-import * as Yup from 'yup';
-
-const resetPasswordSchema = Yup.object().shape({
-	password: Yup.string()
-		.min(8, 'Password must be at least 8 characters')
-		.required('Password is required'),
-	confirmPassword: Yup.string()
-		.oneOf([Yup.ref('password')], 'Passwords must match')
-		.required('Confirm Password is required'),
-});
-
-export { resetPasswordSchema };
