--- src/validations/index.ts
@@ -1,3 +1,0 @@
-export { default as editProfileValidation } from './editProfile.validation';
-export { default as createAccountValidation } from './createAccount.validation';
-export { default as loginFormValidation } from './loginForm.validation';
